Global Exception Handling

+++ src/main/kotlin/com/demo/service/CourseService.kt
@@ -4,12 +4,13 @@ import com.demo.dto.CourseDTO
 import com.demo.entity.Course
 import com.demo.entity.User
 import com.demo.entity.toCourseModel
+import com.demo.exception.ResourceNotFoundException
+import com.demo.exception.ValidationException
 import com.demo.model.CourseModel
 import com.demo.repository.CourseRepository
 import com.demo.repository.UserRepository
-import jakarta.persistence.EntityNotFoundException
 import org.springframework.stereotype.Service
-import java.util.Optional
+import kotlin.jvm.optionals.getOrElse
 
 @Service
 public class CourseService(private val courseRepository: CourseRepository, private val userRepository: UserRepository) {
@@ -19,7 +20,7 @@ public class CourseService(private val courseRepository: CourseRepository, priva
             val courseEntity = Course(null, course.title, course.description, course.author, course.completed)
             val savedCourseEntity = courseRepository.save(courseEntity)
             savedCourseEntity.toCourseModel()
-        }
+        } ?: throw(ValidationException("Invalid Author name provided"))
     }
 
     fun update(
@@ -28,14 +29,14 @@ public class CourseService(private val courseRepository: CourseRepository, priva
     ): CourseModel {
         var course =
             courseRepository.findById(id)
-                .orElseThrow { EntityNotFoundException("Course not found") }
+                .orElseThrow { ResourceNotFoundException("Course not found") }
 
         course.title = courseDTO.title
         course.description = courseDTO.description
         course.completed = courseDTO.completed
 
-        val user: User? = userRepository.findByUsername(courseDTO.author)
-        course.author = user?.username ?: course.author
+        val user = userRepository.findByUsername(courseDTO.author) ?: throw (ValidationException("Invalid author provided."))
+        course.author = user.username
 
         return courseRepository.save(course).toCourseModel()
     }
@@ -45,9 +46,10 @@ public class CourseService(private val courseRepository: CourseRepository, priva
         return listCourseEntities.map { course -> course.toCourseModel() }
     }
 
-    fun fetchOne(id: Long): Optional<CourseModel> {
-        val course: Optional<Course> = courseRepository.findById(id)
-        return course.map { it.toCourseModel() }
+    fun fetchOne(id: Long): CourseModel {
+        return courseRepository.findById(id).map {
+            it.toCourseModel()
+        }.getOrElse { throw ResourceNotFoundException("Course with id: $id does not exist.") }
     }
 
     fun foundOne(id: Long): Boolean {
@@ -55,7 +57,11 @@ public class CourseService(private val courseRepository: CourseRepository, priva
     }
 
     fun delete(id: Long) {
-        courseRepository.deleteById(id)
+        if (foundOne(id)) {
+            courseRepository.deleteById(id)
+        } else {
+            throw ValidationException("Course with id: $id not found.")
+        }
     }
 
     fun searchByAuthorMail(email: String): List<CourseModel>? {
